fixed more routes and controllers bugs
